+++ frontend/src/chatbot.py
@@ -1,8 +1,6 @@
 import streamlit as st
 from langchain_together import ChatTogether
 
-st.title("Basic Chatbot with Langchain and Together.ai")
-
 # Initialize the ChatTogether model
 # Uses TOGETHER_API_KEY environment variable
 llm = ChatTogether(model="meta-llama/Llama-Vision-Free")
@@ -15,32 +13,38 @@ system_prompt = {
 }
 initial_message = {"role": "assistant", "content": "Hello! How can I help you today?"}
 
-# Session state to keep track of conversation
-if "messages" not in st.session_state:
-    st.session_state["messages"] = [initial_message]
 
-# Display previous messages using st.chat_message
-for message in st.session_state["messages"]:
-    with st.chat_message(message["role"]):
-        st.write(message["content"])
+def chatbot():
+    # Session state to keep track of conversation
+    if "messages" not in st.session_state:
+        st.session_state["messages"] = [initial_message]
+    history = st.container(height=850)
 
-# 'Wait' for user input
-if prompt := st.chat_input("You: "):
-    # Display user input, and save it to session state
-    st.chat_message("human").write(prompt)
-    st.session_state["messages"].append({"role": "human", "content": prompt})
+    # Display previous messages using st.chat_message
+    for message in st.session_state["messages"]:
+        history.chat_message(message["role"]).write(message["content"])
 
-    # Generate streamed response
-    full_prompt = [system_prompt] + st.session_state["messages"][-10:]
-    chunked = llm.stream(full_prompt)
+    # 'Wait' for user input
+    if prompt := st.sidebar.chat_input("You: ", key="chat_input"):
+        # Display user input, and save it to session state
+        history.chat_message("human").write(prompt)
+        st.session_state["messages"].append({"role": "human", "content": prompt})
 
-    def wrapped_chunked():
-        for chunk in chunked:
-            if chunk.usage_metadata:
-                print(chunk.usage_metadata)
-            yield chunk
+        # Generate streamed response
+        full_prompt = [system_prompt] + st.session_state["messages"][-10:]
+        chunked = llm.stream(full_prompt)
 
-    response = st.chat_message("assistant").write_stream(wrapped_chunked())
+        def wrapped_chunked():
+            for chunk in chunked:
+                if chunk.usage_metadata:
+                    print(chunk.usage_metadata)
+                yield chunk
 
-    # Save full response to session state
-    st.session_state["messages"].append({"role": "assistant", "content": response})
+        response = history.chat_message("assistant").write_stream(wrapped_chunked())
+
+        # Save full response to session state
+        st.session_state["messages"].append({"role": "assistant", "content": response})
+
+
+if __name__ == "__main__":
+    chatbot()
