+++ shared/setup.py
@@ -5,4 +5,7 @@ setup(
     name="shared",
     version="0.1",
     packages=find_packages(),
+    install_requires=[
+        "pydantic",
+    ],
 )
